requirements.txt
celery
djangorestframework
#git+https://github.com/mukulhase/WebWhatsapp-Wrapper.git
django
numpy
redis
django-cors-headers